Refactor Store Imports and Clean Up Unused Components
- Updated import paths for store-related components to reflect new directory structure.
- Removed obsolete setting store and resource management components to streamline the codebase.
- Adjusted Framework and ResourceTree components to utilize the updated store imports.
- Enhanced type definitions for improved clarity in resource node operations.
- Cleaned up unused imports and optimized component structure for maintainability.
