+++ client/src/src/template/noodle/node.ts
@@ -29,6 +29,8 @@ export const mountNode = async ({ node_key, template_name, mount_params_string }
         node_key = `${leadIP}::${node_key}`
     }
 
+    console.log(node_key, template_name, mount_params_string)
+
     const url = `${API_PREFIX}/mount`
 
     try {
@@ -39,7 +41,6 @@ export const mountNode = async ({ node_key, template_name, mount_params_string }
             },
             body: JSON.stringify({ node_key, template_name, mount_params_string }),
         })
-        console.log(response)
 
         if (!response.ok) {
             throw new Error(`HTTP error! Status: ${response.status}`)
